License fixed and first documentation draft added
